requirements.txt
Flask==2.0.1 
line-bot-sdk==1.20.0 
python-dotenv==0.19.0 
gunicorn==20.1.0 
Werkzeug==2.0.1